UnnecessarySemicolonInspector: code grooming

+++ src/main/java/com/kalessil/phpStorm/phpInspectionsEA/inspectors/codeStyle/UnnecessarySemicolonInspector.java
@@ -2,18 +2,29 @@ package com.kalessil.phpStorm.phpInspectionsEA.inspectors.codeStyle;
 
 import com.intellij.codeInspection.LocalQuickFix;
 import com.intellij.codeInspection.ProblemDescriptor;
-import com.intellij.codeInspection.ProblemHighlightType;
 import com.intellij.codeInspection.ProblemsHolder;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.PsiWhiteSpace;
-import com.intellij.psi.tree.IElementType;
 import com.jetbrains.php.lang.lexer.PhpTokenTypes;
-import com.jetbrains.php.lang.psi.elements.*;
+import com.jetbrains.php.lang.psi.elements.Else;
+import com.jetbrains.php.lang.psi.elements.ElseIf;
+import com.jetbrains.php.lang.psi.elements.If;
+import com.jetbrains.php.lang.psi.elements.Statement;
 import com.kalessil.phpStorm.phpInspectionsEA.openApi.BasePhpElementVisitor;
 import com.kalessil.phpStorm.phpInspectionsEA.openApi.BasePhpInspection;
+import com.kalessil.phpStorm.phpInspectionsEA.utils.OpenapiTypesUtil;
 import org.jetbrains.annotations.NotNull;
+
+/*
+ * This file is part of the Php Inspections (EA Extended) package.
+ *
+ * (c) Vladimir Reznichenko <kalessil@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 public class UnnecessarySemicolonInspector extends BasePhpInspection {
     private static final String message = "Unnecessary semicolon.";
@@ -27,26 +38,19 @@ public class UnnecessarySemicolonInspector extends BasePhpInspection {
     @NotNull
     public PsiElementVisitor buildVisitor(@NotNull final ProblemsHolder holder, boolean isOnTheFly) {
         return new BasePhpElementVisitor() {
-            public void visitPhpStatement(Statement statement) {
-                if (0 == statement.getChildren().length) {
+            @Override
+            public void visitPhpStatement(@NotNull Statement statement) {
+                if (statement.getChildren().length == 0) {
                     final PsiElement parent = statement.getParent();
-                    if (null != parent) {
-                        IElementType declareCandidate = parent.getFirstChild().getNode().getElementType();
-                        if (
-                            PhpTokenTypes.kwDECLARE == declareCandidate ||
-                            parent instanceof DoWhile ||
-                            parent instanceof While ||
-                            parent instanceof For ||
-                            parent instanceof ForeachStatement ||
-                            parent instanceof If ||
-                            parent instanceof ElseIf ||
-                            parent instanceof Else
-                        ) {
-                            return;
-                        }
+                    if (
+                        parent instanceof If || parent instanceof ElseIf || parent instanceof Else ||
+                        OpenapiTypesUtil.isLoop(parent) ||
+                        OpenapiTypesUtil.is(parent.getFirstChild(), PhpTokenTypes.kwDECLARE)
+                    ) {
+                        return;
                     }
 
-                    holder.registerProblem(statement, message, ProblemHighlightType.WEAK_WARNING, new TheLocalFix());
+                    holder.registerProblem(statement, message, new TheLocalFix());
                 }
             }
         };
@@ -68,12 +72,12 @@ public class UnnecessarySemicolonInspector extends BasePhpInspection {
         @Override
         public void applyFix(@NotNull Project project, @NotNull ProblemDescriptor descriptor) {
             final PsiElement expression = descriptor.getPsiElement();
-            if (expression instanceof Statement) {
+            if (expression instanceof Statement && !project.isDisposed()) {
                 /* drop preceding space */
-                if (expression.getPrevSibling() instanceof PsiWhiteSpace) {
-                    expression.getPrevSibling().delete();
+                final PsiElement previous = expression.getPrevSibling();
+                if (previous instanceof PsiWhiteSpace) {
+                    previous.delete();
                 }
-
                 expression.delete();
             }
         }
